Move WordFence to Security namespace
